PIA-634: Create a locator variable for hamburger menu
In LocatorHelper.kt we introduced a new function called findbyContentdesc, as the hamburger menu
do not have any resource ID that is usually use for our previous Login test. Moreover, this function
is created for reusability of oure code.

The findByContentDesc function will search exactly the content-desc string that is declared in a new
created file called SideMenuPageObject.kt, which for out test scenario is called "Open" assigned to
hamburgerMenu variable.

The hamburgerMenu variable will then be used also in a new created file called SideMenuStepObjects.kt
where we invoke a .click method.