PIA-634: Create a locator variable for Settings
In LocatorHelper.kt we introduced a new function called findByText as the Settings does not contain
any resource ID or content description. This function is also created for reusability of our code.

The findByText function will search exactly including case sensitivity the text that is declared in
the filled called SideMenuPageObject.kt which is assigned to settings variable.

The settings variable will be used in a new created function called clickOnSettings in SideMenuStepObjets.kt
where it will invoke a .click() method.