PIA-634: Create a successfulLogout helper
When this helper is called the app will take 3 steps, pressing on the hamburger menu,
selecting logout and confirming that user wants to logout.

This is to simplify and maintain even further our SignOutTests.kt since logging out takes the same steps.
And in the event that the order of the steps will be changed, it can be changed in the ActionHelper file.