PIA-710 - Refactor Sign In test to be data driven (#38)
* PIA-634: Create a locator variable for hamburger menu

In LocatorHelper.kt we introduced a new function called findbyContentdesc, as the hamburger menu
do not have any resource ID that is usually use for our previous Login test. Moreover, this function
is created for reusability of oure code.

The findByContentDesc function will search exactly the content-desc string that is declared in a new
created file called SideMenuPageObject.kt, which for out test scenario is called "Open" assigned to
hamburgerMenu variable.

The hamburgerMenu variable will then be used also in a new created file called SideMenuStepObjects.kt
where we invoke a .click method.

* PIA-634: Create a locator variable for Settings

In LocatorHelper.kt we introduced a new function called findByText as the Settings does not contain
any resource ID or content description. This function is also created for reusability of our code.

The findByText function will search exactly including case sensitivity the text that is declared in
the filled called SideMenuPageObject.kt which is assigned to settings variable.

The settings variable will be used in a new created function called clickOnSettings in SideMenuStepObjets.kt
where it will invoke a .click() method.

* PIA-634-Refactor logout tests to follow best practises

Refactored SideMenuPageObject.kt to SideMenuPageObjects.kt to follow the standard of naming our files.
Instead of using .onClick() function, I expanded the functions for each elements and/or actions.
In this way, we can reuse the functions without getting confuse on what it does. Also, following our
best practises on writing our tests.

SignOutTests.kt contains assertion that checked whether if the default value is showing after logout
after it has been initially change.

* PIA-634: Create a login and settings helper

As there are many steps to take for successful login and going to settings, two function
where created to minimise the repetativeness of it in our SignOuTests.kt

* PIA-634: Refactor method names and create page and step objects

Refactor method names for easier understanding and readability.
Also sepearted objects to specific pages to follow best practises

* PIA-634: Create a successfulLogout helper

When this helper is called the app will take 3 steps, pressing on the hamburger menu,
selecting logout and confirming that user wants to logout.

This is to simplify and maintain even further our SignOutTests.kt since logging out takes the same steps.
And in the event that the order of the steps will be changed, it can be changed in the ActionHelper file.

* PIA-696 - Add helper to navigate back to the Main Screen (#35)

* PIA-634: Create Protocol page object and step object

This is to access protocol settings page so that we can verify whether our system restart the
selection after logout.

* PIA-634: Create assertion for selecting protocol

Refactor and create asserton for selecting protocol after use has successfully logout and then
login of the application.

* PIA-634:Create a test for DIP field for logout

Created DedicatedIPPageObjects and Step objects to access DIP page. This test will check that
the DIP get reset when user logout of the application

* PIA-710 - Made Sign In tests data-driven

* PIA-710 - Remove duplicate dependency

* PIA-710 - Correct typo on defaultTimeOut declaration

---------

Signed-off-by: kp-andrei-madescu <109506442+kp-andrei-madescu@users.noreply.github.com>
Co-authored-by: kp-ruth-o <ruth.ondrada@kape.com>