PIA-634 Refactor test validateDIPSttingsAtLogout
Introduced a action helper called goToSideMenu() as the steps are repetative
(selecting hamburger menu and choosing an option). This action helper can be reuse
as we progress to improve our test coverage further.

Furthermore, in DedicatedStepObjects.kt we combine entering dedicated IP and
clicking on activate button into one function.