PIA-638: Add new tests for protocols with small packets enable and disable
The initial test for this was to test the connectivity of OpenVPN while small packet settings is disabled and
WireGuard small packet settings is enabled. We want to extend the coverage of our tests to make sure that for
the two available protocols have no issue connecting with the vpn whether small packet is enabled or disabled.